Add LineNotify class
